requirements.txt
absl-py==2.4.0
aiobotocore
aiodns
aiohappyeyeballs
aiohttp
aioitertools
aiosignal
alabaster
altair
anaconda-anon-usage
anaconda-auth
anaconda-catalogs
anaconda-cli-base
anaconda-client
anaconda-navigator
anaconda-project
annotated-doc==0.0.4
annotated-types
anyio
appdirs==1.4.4
archspec
argon2-cffi
argon2-cffi-bindings
arrow
astroid
astropy
astropy-iers-data
asttokens
async-lru
asyncssh
atomicwrites==1.4.0
attrs
Automat
autopep8
babel
bcrypt
beautifulsoup4
binaryornot
bitarray==3.8.0
black
bleach
blinker
blis==1.3.3
bokeh
boltons
botocore
Bottleneck
brotlicffi
cached-property==2.0.1
cachetools
catalogue==2.0.10
cattrs
certifi
cffi
chardet
charset-normalizer
chevron==0.14.0
ckzg==2.1.5
click
cloudpathlib==0.23.0
cloudpickle
colorama
colorcet
comm
conda
conda-anaconda-telemetry
conda-anaconda-tos
conda-build
conda-content-trust
conda-libmamba-solver
conda-pack
conda-package-handling
conda-repo-cli
conda_index
conda_package_streaming
confection==0.1.5
constantly
contourpy
cookiecutter
cryptography
cssselect
cycler
cymem==2.0.13
cytoolz==1.1.0
dask
dataclasses-json==0.6.7
datashader
debugpy
decorator
defusedxml
diff-match-patch
dill
distributed
distro
docstring-to-markdown
docutils
easyocr==1.7.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
evalidate
executing
ExifRead==3.5.1
Faker==40.4.0
fastapi==0.128.0
fastjsonschema
filelock
flake8
Flask
flask-cors==6.0.2
flatbuffers==25.12.19
fonttools
frozendict
frozenlist
fsspec
genanki==0.13.1
geographiclib==2.1
geopy==2.4.1
gitdb
GitPython
gmpy2
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-genai==1.64.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet
groq==1.0.0
grpcio==1.78.1
grpcio-status==1.71.2
h11
h5py
HeapDict
hexbytes==1.3.1
holoviews
html5lib
httpcore
httplib2==0.31.2
httpx
httpx-sse==0.4.3
hvplot
hyperlink
idna
ImageHash==4.3.2
ImageIO
imagesize
imbalanced-learn
importlib_metadata
incremental
inflection
iniconfig
intake
intervaltree
ipykernel
ipympl
ipython
ipython_pygments_lexers
ipywidgets
isort
itemadapter
itemloaders
itsdangerous
jaraco.classes
jaraco.context
jaraco.functools
jedi
jellyfish
Jinja2
jiter==0.12.0
jmespath
joblib
json5
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
jupyter
jupyter-console
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
jupyterlab_widgets
keyring
kiwisolver
langchain==1.2.4
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-openai==1.1.7
langchain-text-splitters==1.1.0
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langsmith==0.6.4
lazy_loader
lckr_jupyterlab_variableinspector
libarchive-c
libmambapy
lief
linkify-it-py
llvmlite
lmdb
locket
lsprotocol
lxml
lz4
Markdown
markdown-it-py
MarkupSafe
marshmallow==3.26.2
matplotlib==3.10.6
matplotlib-inline
mccabe
mdit-py-plugins
mdurl
mediapipe==0.10.32
menuinst
mistune
mkl-service==2.5.2
mkl_fft
mkl_random
more-itertools
mpi4py
mpmath
msgpack
multidict
multipledispatch
murmurhash==1.0.15
mypy
mypy_extensions
narwhals
navigator-updater
nbclient
nbconvert
nbformat
nest_asyncio
networkx
ninja==1.13.0
nltk
notebook
notebook_shim
numba
numexpr
numpy
numpydoc
openai==2.15.0
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
openpyxl
orjson==3.11.5
ormsgpack==1.12.1
overrides
packaging
pandas
pandocfilters
panel
param
parsel
parsimonious==0.10.0
parso
partd
pathspec
patsy
pdf2image==1.17.0
pdfminer.six==20251230
pdfplumber==0.11.9
pexpect
pickleshare
pillow
pkce
pkginfo
platformdirs
plotly
pluggy
preshed==3.0.12
prometheus_client
prompt_toolkit
propcache
Protego
proto-plus==1.27.1
protobuf
psutil
ptyprocess
pure_eval
py-cpuinfo
pyarrow
pyasn1
pyasn1_modules
pycares
pyclipper==1.4.0
pycodestyle
pycosat
pycparser
pycryptodome==3.23.0
pyct
pycurl
pydantic
pydantic-settings
pydantic_core
PyDispatcher
pydocstyle
pyerfa
pyflakes
PyGithub
Pygments
PyJWT
pylint
pylint-venv
pyls-spyder==0.4.0
PyMuPDF==1.26.7
PyNaCl
pyodbc
pyOpenSSL
pyparsing
pypdf==6.6.0
pypdfium2==5.5.0
PyQt5==5.15.11
PyQt5_sip
PyQtWebEngine==5.15.7
PySide6==6.9.2
PySocks
pytesseract==0.3.13
pytest
python-bidi==0.6.7
python-dateutil
python-dotenv
python-json-logger
python-lsp-black
python-lsp-jsonrpc
python-lsp-ruff
python-lsp-server
python-multipart==0.0.21
python-slugify
pytokens
pytoolconfig
pytz
pyuca
pyunormalize==17.0.0
pyviz_comms
PyWavelets
pywin32
pywin32-ctypes
pywinpty
PyYAML
pyzmq
QDarkStyle
qstylizer
QtAwesome
qtconsole
QtPy
queuelib
readchar
redis==7.2.0
referencing
regex
requests
requests-file
requests-toolbelt
rfc3339_validator
rfc3986_validator
rich
rlp==4.1.0
roman-numerals-py
rope
rpds-py
rtree
ruamel.yaml
ruamel.yaml.clib
ruamel_yaml_conda
ruff
s3fs
scikit-image
scikit-learn
scipy
Scrapy
seaborn
semver
Send2Trash
service-identity
setuptools==80.9.0
shap==0.50.0
shapely==2.1.2
shellingham
shiboken6==6.9.2
sip
six
slicer==0.0.8
smart_open==7.5.1
smmap
sniffio
snowballstemmer
sortedcontainers
sounddevice==0.5.5
soupsieve
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
Sphinx
sphinxcontrib-applehelp
sphinxcontrib-devhelp
sphinxcontrib-htmlhelp
sphinxcontrib-jsmath
sphinxcontrib-qthelp
sphinxcontrib-serializinghtml
spyder
spyder-kernels
SQLAlchemy
srsly==2.5.2
stack_data
starlette==0.50.0
statsmodels
streamlit
superqt
sympy
tables
tabulate
tblib
tenacity
terminado
text-unidecode
textdistance
thinc==8.3.10
threadpoolctl
three-merge
tifffile
tiktoken==0.12.0
tinycss2
tldextract
toml
tomli
tomlkit
toolz
torch==2.10.0
torchvision==0.25.0
tornado
tqdm
traitlets
truststore
Twisted
twisted-iocpsupport
typer==0.24.1
typer-slim==0.24.0
types-requests==2.32.4.20260107
typing-inspect==0.9.0
typing-inspection
typing_extensions
tzdata
uc-micro-py
ujson
uritemplate==4.2.0
urllib3
uuid_utils==0.13.0
uvicorn==0.40.0
w3lib
wasabi==1.1.3
watchdog
wcwidth
weasel==0.4.3
web3==7.14.1
webencodings
websocket-client
websockets==15.0.1
Werkzeug
whatthepatch
wheel==0.45.1
widgetsnbextension
win_inet_pton
winloop
wrapt
xarray
xlwings
xxhash==3.6.0
xyzservices
yapf
yarl
zict
zipp
zope.interface
zstandard
gunicorn